Use Perl's savestack mechanism to clear TCC mutex on exception
One flaw of the prior commit was that it would fail to clear the
TCC-protecting mutex if there was an exception thrown between
pthread_mutex_lock() and pthread_mutex_unlock().

With this change (view it with "git diff -w"!), we're now using Perl's
save stack to execute a callback (which unlocks the mutex) when we exit
the "current pseudoblock". We also add an ENTER/LEAVE pair around the
code section that needs to be protected, which creates such a pseudo
block.

This appears to work.

Except that there seems to be another thread-unsafe piece of code
somewhere. With the amended (much more aggressive) test script, I get
this on some runs:

  ~/perl/C-Blocks$ perl -Mblib test_script_for_tcc_thread_segfaults.pl
  Spawned 3 threads at test_script_for_tcc_thread_segfaults.pl line 25.
  Thread 1 of 3 remaining is joinable at test_script_for_tcc_thread_segfaults.pl line 31.
  Joined thread at test_script_for_tcc_thread_segfaults.pl line 33.
  Thread 1 of 2 remaining is joinable at test_script_for_tcc_thread_segfaults.pl line 31.
  Joined thread at test_script_for_tcc_thread_segfaults.pl line 33.
  Thread 1 of 1 remaining is joinable at test_script_for_tcc_thread_segfaults.pl line 31.
  Joined thread at test_script_for_tcc_thread_segfaults.pl line 33.
  Got to end of script at test_script_for_tcc_thread_segfaults.pl line 41.
  Segmentation fault (core dumped)

Inspecting the core file, I see: (This is not a debugging perl, but
that's fine for the conclusion here.)

  ~/perl/C-Blocks$ gdb perl -c core
  GNU gdb (Ubuntu 7.11.1-0ubuntu1~16.04) 7.11.1
  ...
  Reading symbols from perl...(no debugging symbols found)...done.
  [New LWP 6826]
  [Thread debugging using libthread_db enabled]
  Using host libthread_db library "/lib/x86_64-linux-gnu/libthread_db.so.1".
  Core was generated by `perl -Mblib test_script_for_tcc_thread_segfaults.pl'.
  Program terminated with signal SIGSEGV, Segmentation fault.
  #0  0x000000000041ef88 in Perl_Slab_Free ()
  (gdb) bt
  #0  0x000000000041ef88 in Perl_Slab_Free ()
  #1  0x000000000041fc70 in Perl_op_free ()
  #2  0x000000000043eeb8 in perl_destruct ()
  #3  0x000000000041cb83 in main ()
  (gdb) quit

So I think the segfault is happening during global destruction now.
There's another failure I'm managing to produce, which is:

  Thread 1 of 3 remaining is joinable at test_script_for_tcc_thread_segfaults.pl line 31.
  *** Error in `perl': double free or corruption (out): 0x00000000017abd10 ***
  ======= Backtrace: =========
  /lib/x86_64-linux-gnu/libc.so.6(+0x777e5)[0x7f35ffca77e5]
  /lib/x86_64-linux-gnu/libc.so.6(+0x7fe0a)[0x7f35ffcafe0a]
  /lib/x86_64-linux-gnu/libc.so.6(cfree+0x4c)[0x7f35ffcb398c]
  perl(Perl_reentrant_free+0x13)[0x4a6243]
  perl(perl_destruct+0x13b0)[0x440120]
  /usr/lib/x86_64-linux-gnu/perl/5.22/auto/threads/threads.so(+0x329a)[0x7f35fec9029a]
  /usr/lib/x86_64-linux-gnu/perl/5.22/auto/threads/threads.so(+0x514a)[0x7f35fec9214a]
  perl(Perl_pp_entersub+0x48a)[0x4bd3ca]
  perl(Perl_runops_standard+0x16)[0x4b6276]
  perl(perl_run+0x2f9)[0x443c59]
  perl(main+0x12b)[0x41cbbb]
  /lib/x86_64-linux-gnu/libc.so.6(__libc_start_main+0xf0)[0x7f35ffc50830]
  perl(_start+0x29)[0x41cbf9]

My guess is that there's something in C::Blocks that holds on to what
amounts to shared state which thus gets freed at thread exit already?
That can then trip up either because another worker thread or the main
thread end up needing to use said state or because it's double-freed at
another thread or global exit.

Speculation: OP trees are generally shared between threads and thus
the custom OP that's generated for the clex block and its dependent
data should only be freed once (or never depending on how it was
implemented - haven't checked yet). So while this is the most obvious
state to stick around, I somehow doubt that's it.

Needs more digging, but maybe the above gives David a hunch already.