Nasty hack to port the XS::TCC automatic XS wrapping hack
This is not to be merged as is.

This is a terrible hack to support a "cfun" block which, like a "csub"
block generates a Perl-visible XSUB. Like XS::TCC, "cfun" will actually
generate the stack munging/type mapping code for the user based on a
typemap (right now just the core perl typemap) and a C function
signature.

The syntax should be something like this:

  cfun double square(double x) {
    return x*x;
  }

But because I couldn't spend the time to make the perl lexer/parser
parse the C function signature, I did it the hacky way: The first line
of the code needs to be a bare C function signature. This is intended
just for the proof of principle! So this now works:

  cfun square {
    double foo (double x)
      return x*x;
  }

  warn square(3);
  # warns 9